pyproject.toml
[tool.poetry]
name = "audio2txt"
version = "3.0.0-alpha.1"
description = "Next Generation AI Audio Transcription Platform"
authors = ["Audio2txt Team"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
faster-whisper = "^1.0.0"
pyannote-audio = "^3.1.0"
librosa = "^0.10.0"
torch = "^2.1.0"
transformers = "^4.36.0"
opencc-python-reimplemented = "^0.1.7"
redis = "^5.0.0"
celery = {extras = ["redis"], version = "^5.3.0"}
reportlab = "^4.0.0"
structlog = "^24.0.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
click = "^8.1.0"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
audio2txt = "apps.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["packages/core/tests"]
addopts = ["--cov=packages/core/audio2txt"]
